Split up the ExcerptInterface
