+++ packages/content/src/Infrastructure/Sulu/Admin/ContentViewBuilderFactory.php
@@ -27,6 +27,7 @@ use Sulu\Content\Application\ContentMetadataInspector\ContentMetadataInspectorIn
 use Sulu\Content\Domain\Model\ExcerptInterface;
 use Sulu\Content\Domain\Model\SeoInterface;
 use Sulu\Content\Domain\Model\ShadowInterface;
+use Sulu\Content\Domain\Model\TaxonomyInterface;
 use Sulu\Content\Domain\Model\TemplateInterface;
 use Sulu\Content\Domain\Model\WorkflowInterface;
 
@@ -189,7 +190,9 @@ class ContentViewBuilderFactory implements ContentViewBuilderFactoryInterface
                 );
             }
 
-            if (\is_subclass_of($dimensionContentClass, ExcerptInterface::class)) {
+            if (\is_subclass_of($dimensionContentClass, ExcerptInterface::class)
+                || \is_subclass_of($dimensionContentClass, TaxonomyInterface::class)
+            ) {
                 $views[] = $this->createExcerptFormView(
                     $editParentView,
                     $previewEnabled,
@@ -282,11 +285,14 @@ class ContentViewBuilderFactory implements ContentViewBuilderFactoryInterface
             }
         }
 
+        $hasExcerpt = \is_subclass_of($dimensionContentClass, ExcerptInterface::class);
+        $tabTitle = $hasExcerpt ? 'sulu_content.excerpt' : 'sulu_content.taxonomies';
+
         return $this->createFormViewBuilder($parentView . '.excerpt', '/excerpt', $previewEnabled)
             ->addMetadataRequestParameters(['forms' => $forms])
             ->setResourceKey($resourceKey)
             ->setFormKey('content_excerpt')
-            ->setTabTitle('sulu_content.excerpt')
+            ->setTabTitle($tabTitle)
             ->setTitleVisible(true)
             ->addToolbarActions(\array_values($toolbarActions))
             ->setTabOrder(40)
