+++ packages/content/translations/admin.de.json
@@ -1,6 +1,7 @@
 {
     "sulu_content.seo": "SEO",
     "sulu_content.excerpt": "Auszug & Taxonomien",
+    "sulu_content.taxonomies": "Taxonomien",
     "sulu_content.content": "Inhalt",
     "sulu_content.text_for_more_link": "Text für weiterführenden Link",
     "sulu_content.icon": "Icon",
