+++ packages/content/translations/admin.en.json
@@ -1,6 +1,7 @@
 {
     "sulu_content.seo": "SEO",
     "sulu_content.excerpt": "Excerpt & Taxonomies",
+    "sulu_content.taxonomies": "Taxonomies",
     "sulu_content.content": "Content",
     "sulu_content.text_for_more_link": "Text for more link",
     "sulu_content.icon": "Icon",
